requirements.txt
absl-py==1.0.0
biopython==1.79
chex==0.0.7
dm-haiku==0.0.10
dm-tree==0.1.8
docker==5.0.0
immutabledict==2.0.0
jax==0.4.14
matplotlib==3.8.3
ml-collections==0.1.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
tensorflow-cpu==2.13.0
